+++ pkg/controller/nativeResourceWorker_test.go
@@ -24,7 +24,17 @@ var _ = Describe("Routes", func() {
 	var mockCtlr *mockController
 	BeforeEach(func() {
 		mockCtlr = newMockController()
-		mockCtlr.multiClusterHandler = NewClusterHandler("")
+		params := Params{
+			MultiClusterMode: PrimaryCIS,
+			Agent: &Agent{
+				PostManager: &PostManager{
+					PrimaryClusterHealthProbeParams: PrimaryClusterHealthProbeParams{
+						statusRunning: true,
+					},
+				},
+			},
+		}
+		mockCtlr.multiClusterHandler = NewClusterHandler("", params.MultiClusterMode, &params.Agent.PrimaryClusterHealthProbeParams)
 		go mockCtlr.multiClusterHandler.ResourceEventWatcher()
 		// Handles the resource status updates
 		go mockCtlr.multiClusterHandler.ResourceStatusUpdater()
@@ -1933,7 +1943,17 @@ var _ = Describe("With NamespaceLabel parameter in deployment", func() {
 	var mockCtlr *mockController
 	BeforeEach(func() {
 		mockCtlr = newMockController()
-		mockCtlr.multiClusterHandler = NewClusterHandler("")
+		params := Params{
+			MultiClusterMode: PrimaryCIS,
+			Agent: &Agent{
+				PostManager: &PostManager{
+					PrimaryClusterHealthProbeParams: PrimaryClusterHealthProbeParams{
+						statusRunning: true,
+					},
+				},
+			},
+		}
+		mockCtlr.multiClusterHandler = NewClusterHandler("", params.MultiClusterMode, &params.Agent.PrimaryClusterHealthProbeParams)
 		go mockCtlr.multiClusterHandler.ResourceEventWatcher()
 		// Handles the resource status updates
 		go mockCtlr.multiClusterHandler.ResourceStatusUpdater()
@@ -2022,7 +2042,17 @@ var _ = Describe("Without NamespaceLabel parameter in deployment", func() {
 	var mockCtlr *mockController
 	BeforeEach(func() {
 		mockCtlr = newMockController()
-		mockCtlr.multiClusterHandler = NewClusterHandler("")
+		params := Params{
+			MultiClusterMode: PrimaryCIS,
+			Agent: &Agent{
+				PostManager: &PostManager{
+					PrimaryClusterHealthProbeParams: PrimaryClusterHealthProbeParams{
+						statusRunning: true,
+					},
+				},
+			},
+		}
+		mockCtlr.multiClusterHandler = NewClusterHandler("", params.MultiClusterMode, &params.Agent.PrimaryClusterHealthProbeParams)
 		go mockCtlr.multiClusterHandler.ResourceEventWatcher()
 		// Handles the resource status updates
 		go mockCtlr.multiClusterHandler.ResourceStatusUpdater()
@@ -2072,7 +2102,17 @@ var _ = Describe("Multi Cluster with Routes", func() {
 
 	BeforeEach(func() {
 		mockCtlr = newMockController()
-		mockCtlr.multiClusterHandler = NewClusterHandler("")
+		params := Params{
+			MultiClusterMode: PrimaryCIS,
+			Agent: &Agent{
+				PostManager: &PostManager{
+					PrimaryClusterHealthProbeParams: PrimaryClusterHealthProbeParams{
+						statusRunning: true,
+					},
+				},
+			},
+		}
+		mockCtlr.multiClusterHandler = NewClusterHandler("", params.MultiClusterMode, &params.Agent.PrimaryClusterHealthProbeParams)
 		go mockCtlr.multiClusterHandler.ResourceEventWatcher()
 		// Handles the resource status updates
 		go mockCtlr.multiClusterHandler.ResourceStatusUpdater()
@@ -2375,7 +2415,17 @@ var _ = Describe("Multi Cluster with CRD", func() {
 
 	BeforeEach(func() {
 		mockCtlr = newMockController()
-		mockCtlr.multiClusterHandler = NewClusterHandler("")
+		params := Params{
+			MultiClusterMode: PrimaryCIS,
+			Agent: &Agent{
+				PostManager: &PostManager{
+					PrimaryClusterHealthProbeParams: PrimaryClusterHealthProbeParams{
+						statusRunning: true,
+					},
+				},
+			},
+		}
+		mockCtlr.multiClusterHandler = NewClusterHandler("", params.MultiClusterMode, &params.Agent.PrimaryClusterHealthProbeParams)
 		go mockCtlr.multiClusterHandler.ResourceEventWatcher()
 		// Handles the resource status updates
 		go mockCtlr.multiClusterHandler.ResourceStatusUpdater()