+++ docs/RELEASE-NOTES.rst
@@ -13,6 +13,8 @@ Added Functionality
 Bug Fixes
 ````````````
 * `Issue 3679 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3679>`_: Certificate, CA chain, and private key shown in debug logs
+* `Issue 3726 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3726>`_: VS in default multi cluster mode not working
+* `Issue 3724 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3724>`_: Transport server status “ERROR” on the secondary cluster
 
 2.19.0
 -------------